Trabajo de grado/Backend/Api/Serializers.py
# backend/api/serializers.py
from rest_framework import serializers
from Models.py import AudioFile, Instrument

class AudioFileSerializer(serializers.ModelSerializer):
    class Meta:
        model = AudioFile
        fields = '__all__'

class InstrumentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Instrument
        fields = '__all__'
