Trabajo de grado/Backend/requirements.txt
Django==3.2.4
djangorestframework==3.12.4
django-cors-headers==3.7.0
Pillow==8.2.0
librosa==0.8.1
scikit-learn==0.24.2
numpy==1.20.3
pandas==1.2.4

